mcp/aws-prometheus-mcp/lambda/requirements.txt
PyJWT==2.8.0
cryptography==41.0.7
